Change Source0 URL
Current URL redirects to https://github.com/voxpupuli/puppet-corosync
which is causing issues in tarname detection as repo name changes
from puppetlab-corosync to puppet-corosync.

This patch changes url to github.com/puppet-community/puppet-corosync
which redirects to the same module but keeps the same repo name.

With this change upstream_name is not longer required.

Change-Id: Id8d20af3a572d7bf4619d57c29960ac7951cd130